Move events page to js folder
