+++ static/js/page/events/_config.scss
