+++ static/js/page/events/dropdown/styles.scss
