+++ static/js/page/events/event/styles.scss
@@ -1,5 +1,5 @@
-@import '../../../../css/_config';
-@import '../../../../css/_utils';
+@import '../../../../css/config';
+@import '../../../../css/utils';
 
 .event {
   padding: 20px;
