+++ static/js/page/events/event/view.twig
