+++ static/js/page/events/events-list.js
