+++ static/js/page/events/events-list/EventsList.js
@@ -2,7 +2,7 @@ require('./styles.scss');
 
 var $ = require('jquery');
 var template = require('./view.twig');
-var emitter = require('../../../utils/emitter');
+var emitter = require('../../../util/emitter');
 var EVENTS = require('../events-list');
 
 /**
