+++ static/js/page/events/events-list/styles.scss
