+++ static/js/page/events/events-list/view.twig
