+++ static/js/page/events/filter-panel/styles.scss
