+++ static/js/page/events/filter-panel/view.twig
