+++ static/js/page/events/index.scss
@@ -1,8 +1,8 @@
 @import 'config';
-@import '../../../css/_config';
-@import '../../../css/_utils';
-@import '../../../css/com/_bits/_header';
-@import '../../../css/com/_bits/_table';
+@import '../../../css/config';
+@import '../../../css/utils';
+@import '../../../css/com/_bits/header';
+@import '../../../css/com/_bits/table';
 
 body {
   overflow-x: hidden;
