+++ static/js/page/events/map/marker-icon-highlighted.png
