+++ static/js/page/events/map/marker-icon.png
