+++ static/js/page/events/map/styles.js
