+++ static/js/page/events/switcher/index.js
