+++ static/js/page/events/switcher/styles.scss
