+++ static/js/page/events/switcher/view.twig
